+++ src/main/java/com/springboot/springboot/model/UpdateHouseRequest.java
@@ -2,13 +2,13 @@ package com.springboot.springboot.model;
 
 public class UpdateHouseRequest {
 
-    public double HousePrice;
-    public double LotArea;
+    public float HousePrice;
+    public float LotArea;
     public String Street;
     public String SaleCondtion;
     public int YearBuilt;
 
-    public UpdateHouseRequest(double housePrice, double lotArea, String street, String saleCondtion, int yearBuilt) {
+    public UpdateHouseRequest(float housePrice, float lotArea, String street, String saleCondtion, int yearBuilt) {
         HousePrice = housePrice;
         LotArea = lotArea;
         Street = street;
